pyproject.toml
[tool.poetry]
name = "bfrl"
version = "0.1.0"
description = ""
authors = ["MeadMaker"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tcod = "^14.0.0"
pyyaml = "^6.0"
pygame = {git = "https://github.com/pygame/pygame.git"}

[tool.pyright]
venvPath = "/Users/pfass/Library/Caches/pypoetry/virtualenvs"
venv = "bfrl-93mGb6aN-py3.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
